addad possibility to specify version code to promote
when promoting an app from one track to the other, previously the latest
release on the source track was promoted.
added a possibility to specify which version code to promote from the
source track.
if no version code is specified, the latest version on source track is
used
if a version code is specified, the apk of that version code will be
promoted.
if there is no apk with that version code on the source track, the task
will fail