updated pr check across all targeted branch (#482)
[#AB2331316](https://mseng.visualstudio.com/AzureDevOps/_workitems/edit/2331316)

**Description**: <Describe your changes here>
Build check is not automatically getting triggered upon PR raised. It is
getting triggered upon /azp run but the status is not getting updated.
Followed the same standard as defined in another
[repo](https://dev.azure.com/mseng/PipelineTools/_apps/hub/ms.vss-build-web.ci-designer-hub?pipelineId=7502&branch=master&nonce=5NogWbHEshu7Tjp6%2BGJLQA%3D%3D).

And, we have enabled the "**report build status**" option as shown in
the image below? This is for the pipeline: [Pipelines - Runs for
microsoft.google-play-vsts-extension](https://mseng.visualstudio.com/PipelineTools/PipelineTools%20Team/_build?definitionId=21513&_a=summary).
We can find it by going to View -> then selecting the three vertical
dots -> Triggers -> and then YAML.

**Documentation changes required:** (Y/N) <Please mark if documentation
changes are required>

**Added unit tests:** (Y/N) <Please mark if unit tests were added or
updated according changes>

**Attached related issue:** (Y/N) <Please add link to related issue
here>

**Checklist**:
- [ ] Task version was bumped - please check
[instruction](https://github.com/microsoft/google-play-vsts-extension/tree/master/docs/taskversionbumping.md)
how to do it
- [ ] Checked that applied changes work as expected

Co-authored-by: azure-pipelines-bot <azure-pipelines-bot@microsoft.com>